requirements.txt
Flask
gunicorn
Pillow
requests
python-dotenv